simple_vision/requirements.txt
httpx
python-dotenv
Pillow
